Create asynchrone event to manage the user input.
The api call is made only after 400ms without any input from the user. (to prevent useless api calls).
The api call is not made if the input is less than 3 char long.
The api call is not made if nothing has changed inside the input.
If the api call is made but a new api call is made before the api has returned the answer, the api call is canceled and a new one is made.