Add a new NavitiaService file to separate responsabilities between the client, the service and the ressource.
The resource exposes endpoint to allow the frontend to get or post data. It calls the service methods.
The service will manage all api calls. It will process the results and make new api calls if needed. It uses the client to make the api calls.
The client it used to make the calls (literally)