Save search results when navigating between configuration page, results page and details page.
User can now go back to the journey proposal list without having to reset the trip all over again.
If the user goes back, the api call to the backend is not triggered again. The journey list is saved and immediately shown.
It the user decides to search for a new trip, the journey proposals list is erased and the search is triggered again with new parameters